Implement functionality to enumerate multiple domains and some minor fixes
This commit adds a new function that can read domain names from a file or from sys.stdin. The function also checks and normalizes the input to make sure that the domains are valid and in the correct format.
This change streamlines the input process and makes it easier for users to provide a list of domains for the program to work with.
Also applied missing logic for args.plain and changed argument switch for target domain to `-d` which was previously `-t` to avoid conflict with upcoming implimantation of threads option.